@@ -14,23 +14,23 @@
   "private": true,
   "sideEffects": false,
   "dependencies": {
-    "@angular/common": "^13.1.1",
-    "@angular/compiler": "^13.1.1",
-    "@angular/core": "^13.1.1",
-    "@angular/elements": "^13.1.1",
-    "@angular/forms": "^13.1.1",
-    "@angular/localize": "^13.1.1",
-    "@angular/platform-browser": "^13.1.1",
-    "@angular/platform-browser-dynamic": "^13.1.1",
-    "@angular/router": "^13.1.1",
-    "@angular/service-worker": "^13.1.1",
+    "@angular/common": "^14.1.3",
+    "@angular/compiler": "^14.1.3",
+    "@angular/core": "^14.1.3",
+    "@angular/elements": "^14.1.3",
+    "@angular/forms": "^14.1.3",
+    "@angular/localize": "^14.1.3",
+    "@angular/platform-browser": "^14.1.3",
+    "@angular/platform-browser-dynamic": "^14.1.3",
+    "@angular/router": "^14.1.3",
+    "@angular/service-worker": "^14.1.3",
     "@fortawesome/angular-fontawesome": "^0.10.0",
     "@fortawesome/fontawesome-svg-core": "^6.0.0",
     "@fortawesome/free-brands-svg-icons": "^6.0.0",
     "@fortawesome/free-regular-svg-icons": "^6.0.0",
     "@fortawesome/free-solid-svg-icons": "^6.0.0",
-    "@ng-bootstrap/ng-bootstrap": "^12.0.0",
-    "bootstrap": "5.1.3",
+    "@ng-bootstrap/ng-bootstrap": "^13.0.0",
+    "bootstrap": "^5.2.0",
     "lodash.clonedeep": "^4.5.0",
     "ngx-clipboard": "^15.0.1",
     "ngx-sport": "^1.0.23",
@@ -41,8 +41,8 @@
   "devDependencies": {
     "@angular-devkit/build-angular": "^13.1.2",
     "@angular/cli": "^13.1.2",
-    "@angular/compiler-cli": "^13.1.1",
-    "@angular/language-service": "^13.1.1",
+    "@angular/compiler-cli": "^14.1.3",
+    "@angular/language-service": "^14.1.3",
     "@types/jasmine": "^3.10.2",
     "@types/jasminewd2": "^2.0.10",
     "@types/lodash.clonedeep": "^4.5.6",
@@ -56,7 +56,7 @@
     "karma-jasmine": "^4.0.0",
     "protractor": "^7.0.0",
     "ts-node": "^9.0.0",
-    "typescript": "~4.4.4",
+    "typescript": "~4.7.4",
     "typescript-eslint": "0.0.1-alpha.0",
     "webpack-bundle-analyzer": "^4.5.0",
     "webpack-dev-server": "^4.7.1"
